+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ### New features
 
-- Directed acyclic graph (DAG) pipelines - where the output of one pipeline can feed into another - are now available using the `maestroOutputs` and `maestroInputs` tags. Pipelines that input into a downstream pipeline should use the `maestroOutputs` tag. Pipelines that receive input from an upstream pipeline should use the `maestroInputs` tag.
+- Directed acyclic graph (DAG) pipelines - where the output of one pipeline can feed into another - are now available using the `maestroOutputs` and `maestroInputs` tags. Pipelines that input into a downstream pipeline should use the `maestroOutputs` tag. Pipelines that receive input from an upstream pipeline should use the `maestroInputs` tag (#98).
 
 - New function `show_network` for visualizing the connections between pipelines that are connected in a DAG.
 
@@ -17,6 +17,8 @@
 ### Bug fixes
 
 - Fixed issue with `suggest_orch_frequency` when using different styles of frequency (e.g., 1 day vs. daily) in a single schedule.
+
+- Fixed issue where pipeline sourcing failures were appearing as successful runs in status outputs.
 
 # maestro 0.3.0